+++ src/logger/console-logger.ts
@@ -1,5 +1,6 @@
 'use strict'
 
+import LogUpdate from 'log-update'
 import { tap } from '@supercharge/goodies'
 import { LoggerContract } from './logger-contract'
 
@@ -7,26 +8,50 @@ export class ConsoleLogger implements LoggerContract {
   /**
    * Log the given `message` to the output console using `console.log`.
    *
-   * @param message
+   * @param {String} message
    *
-   * @returns {ConsoleOutput}
+   * @returns {ConsoleLogger}
    */
-  log (message: string): this {
+  log (message: any, ...optionalParams: any[]): this {
     return tap(this, () => {
-      console.log(message)
+      console.log(message, ...optionalParams)
     })
   }
 
   /**
    * Log the given `message` to the error console using `console.error`.
    *
-   * @param message
+   * @param {String} message
    *
-   * @returns {ConsoleOutput}
+   * @returns {ConsoleLogger}
    */
-  logError (message: string): this {
+  logError (message: any, ...optionalParams: any[]): this {
     return tap(this, () => {
-      console.error(message)
+      console.error(message, ...optionalParams)
+    })
+  }
+
+  /**
+   * Log the given `message` to the terminal by overwriting the previous message.
+   *
+   * @param {String} message
+   *
+   * @returns {ConsoleLogger}
+   */
+  logUpdate (message: string): this {
+    return tap(this, () => {
+      LogUpdate(message)
+    })
+  }
+
+  /**
+    * Persist the log message previosly written to the terminal using `logUpdate`.
+    *
+    * @returns {ConsoleLogger}
+    */
+  logUpdateDone (): this {
+    return tap(this, () => {
+      LogUpdate.done()
     })
   }
 }